+++ src/ConsumableItem.php
@@ -37,6 +37,7 @@ use Glpi\DBAL\QueryExpression;
 use Glpi\Application\View\TemplateRenderer;
 use Glpi\DBAL\QueryFunction;
 use Glpi\Features\AssetImage;
+use Glpi\Features\AssignableAsset;
 
 //!  ConsumableItem Class
 /**
@@ -47,6 +48,7 @@ use Glpi\Features\AssetImage;
 class ConsumableItem extends CommonDBTM
 {
     use AssetImage;
+    use AssignableAsset;
 
    // From CommonDBTM
     protected static $forward_entity_to = ['Consumable', 'Infocom'];
@@ -460,16 +462,6 @@ class ConsumableItem extends CommonDBTM
         $options['entities_id'] = $this->getEntityID();
         $options['items']       = [$item];
         NotificationEvent::debugEvent($this, $options);
-    }
-
-
-    public function canUpdateItem()
-    {
-
-        if (!$this->checkEntity(true)) { //check entities recursively
-            return false;
-        }
-        return true;
     }
 
     public function showForm($ID, array $options = [])