+++ src/NetworkName.php
@@ -971,4 +971,13 @@ class NetworkName extends FQDNLabel
         }
         return '';
     }
+
+    public function getRights($interface = 'central')
+    {
+        $rights = parent::getRights($interface);
+        // Rename READ and UPDATE right labels to match other assets
+        $rights[READ] = __('View all');
+        $rights[UPDATE] = __('Update all');
+        return $rights;
+    }
 }