+++ src/Printer.php
@@ -46,6 +46,7 @@ class Printer extends CommonDBTM
     use Glpi\Features\Clonable;
     use Glpi\Features\Inventoriable;
     use Glpi\Features\State;
+    use Glpi\Features\AssignableAsset;
 
    // From CommonDBTM
     public $dohistory                   = true;
@@ -808,7 +809,6 @@ class Printer extends CommonDBTM
     {
         return $this->restore(["id" => $ID]);
     }
-
 
     public static function getIcon()
     {