+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The present file will list all changes made to the project; according to the
 - `Deny login` authorization rule action to deny login for a user, but not prevent the import/existence of the user in GLPI.
 - Directly capture screenshots or screen recordings from the "Add a document" form in tickets.
 - With a clean install, dashboards now show fake/placeholder data by default with a message indicating you are viewing demonstration data and a button to disable it.
+- Assets that can be assigned to users/groups have new "View assigned" and "Update assigned" rights which give read/update access to users and groups assigned to the asset.
 
 ### Changed
 - ITIL Objects can now be linked to any other ITIL Objects similar to the previous Ticket/Ticket links.